Moved managers to correct namespace
